+++ src/redux/modules/index.jsx
@@ -1,14 +1,12 @@
 import { combineReducers } from 'redux';
 import b1 from './b1';
 import h2A from './h2A';
-import h3 from './h3';
-import h31 from './h31';
 import i from './i';
 import i1 from './i1';
 import g from './g';
 import p1 from './p1';
 import ox2 from './ox2';
-import h1 from './h1';
+import h1h3h31 from './h1h3h31';
 import f from './f';
 import f2 from './f2';
 
@@ -17,14 +15,12 @@ const rootReducer = combineReducers({
   // inspection,
   b1,
   h2A,
-  h3,
-  h31,
+  h1h3h31,
   i,
   i1,
   g,
   p1,
   ox2,
-  h1,
   f,
   f2,
 });
